feature : 계약관리 검색 조건 함수 추가
계약 정보 조회시 검색 조건에 따라 and 조건을 통해 계약 정보를 조회 할 수 있음

close #4

+++ src/main/java/com/clover/salad/contract/command/dto/ContractUploadResponseDTO.java
@@ -13,8 +13,8 @@ public class ContractUploadResponseDTO {
 	private int contractId;
 	private String message;
 
-	public ContractUploadResponseDTO(Long id, String message) {
-		this.contractId = (id != null) ? id.intValue() : -1; // null 처리
+	public ContractUploadResponseDTO(Integer id, String message) {
+		this.contractId = (id != null) ? id : -1; // null 처리
 		this.message = message;
 	}
 }
